lance-quanropi-ogummaa/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-genai>=0.7.0
google-cloud-aiplatform>=1.50.0
google-cloud-storage>=2.16.0
lancedb>=0.9.0
pyarrow>=15.0.0
pydantic>=2.7.0
python-dotenv>=1.0.1
